+++ exercises/practice/sum-of-multiples/sum_of_multiples.zig
@@ -1,7 +1,7 @@
 const std = @import("std");
 const mem = std.mem;
 
-pub fn sum(allocator: mem.Allocator, factors: []const usize, limit: usize) !usize {
+pub fn sum(allocator: mem.Allocator, factors: []const u32, limit: u32) !u32 {
     _ = allocator;
     _ = factors;
     _ = limit;
